requirements.txt
FFlask==2.0.2
Werkzeug==2.0.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
Flask-Migrate==3.1.0